Price-related changes: ListView price filter now filters prices above-or-equal-to the searched price. [fix] The default price in the category table is now a float, not an int. The preset categories now include a default price.

+++ class/citruscategory.class.php
@@ -19,9 +19,9 @@ class CitrusCategory extends SeedObject
         global $langs;
         $this->db = $db;
         $this->fields = array(
-            'code'           => array('type' => 'string', 'length' => 255, 'index' => true),
-            'label'          => array('type' => 'string', 'length' => 1024),
-            'default_price' => array('type' => 'integer'),
+            'code'          => array('type' => 'string', 'length' => 255, 'index' => true),
+            'label'         => array('type' => 'string', 'length' => 1024),
+            'default_price' => array('type' => 'float'),
             'active'        => array('type' => 'integer')
         );
         $this->init();
